bug fix : scalar upload error

+++ dolphindb_csharpapi/data/AbstractScalar.cs
@@ -25,7 +25,7 @@ namespace dolphindb.data
         public void write(ExtendedDataOutput @out)
         {
             int flag = ((int)DATA_FORM.DF_SCALAR << 8) + (int)getDataType();
-            @out.writeInt(flag);
+            @out.writeShort(flag);
             writeScalarToOutputStream(@out);
         }
